requirements.txt
networkx==3.1
numba==0.56.4
numpy==1.23.5
ogb==1.3.6
performer-pytorch==1.1.4
pyarrow==11.0.0
pytorch-lightning==1.9.0
requests==2.28.2
tensorboardX==2.5.1
torchmetrics==0.11.0
typer==0.7.0
wandb==0.15.1
yacs==0.1.8
# For plotting
matplotlib==3.7.1
seaborn==0.12.2
# For download pre-trained models
